% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_glmm.R
\name{.glmm_connection}
\alias{.glmm_connection}
\title{Connection form A at one parameter point}
\usage{
.glmm_connection(G_FF_diag, G_BF)
}
\arguments{
\item{G_FF_diag}{J-vector returned by \code{\link[=.glmm_G_FF]{.glmm_G_FF()}}.}

\item{G_BF}{J x 2 matrix returned by \code{\link[=.glmm_G_BF]{.glmm_G_BF()}}.}
}
\value{
J x 2 matrix.
}
\description{
A = -G_FF^{-1} G_BF^T   \link{J x 2}
}
\details{
The horizontal lift of a base motion dtheta = (dmu, dsigma) is
dalpha = A \%*\% dtheta.

A\link{j, 1} =  1 / (sigma^2 * G_FF\link{j})
A\link{j, 2} =  2*(alpha_j - mu) / (sigma^3 * G_FF\link{j})
}
\keyword{internal}
