% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_glmm.R
\name{.glmm_curvature}
\alias{.glmm_curvature}
\title{Analytic curvature of the GLMM connection}
\usage{
.glmm_curvature(G_FF_diag, sigma)
}
\arguments{
\item{G_FF_diag}{J-vector.}

\item{sigma}{Scalar.}
}
\value{
J-vector.
}
\description{
For the 2D base (mu, sigma), the curvature collapses to a J-vector:
}
\details{
F\link{j} = d_mu A\link{j,sigma} - d_sigma A\link{j,mu}
= -2 / (sigma^5 * G_FF\link{j}^2)

F\link{j} < 0 for all j: the connection is contractive (not rotational) for
this model.  |F\link{j}| grows as the prior dominates (small data, large sigma).
}
\keyword{internal}
