% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_glmm.R
\name{.glmm_G_BF}
\alias{.glmm_G_BF}
\title{Off-diagonal block G_BF at one parameter point}
\usage{
.glmm_G_BF(sigma, mu, alpha)
}
\arguments{
\item{sigma}{Scalar.}

\item{mu}{Scalar.}

\item{alpha}{J-vector.}
}
\value{
J x 2 matrix.
}
\description{
G_BF[j, 1] = -1/sigma^2              (base direction: mu)
G_BF[j, 2] = -2*(alpha_j - mu)/sigma^3   (base direction: sigma)
}
\keyword{internal}
