% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_glmm.R
\name{.glmm_G_FF}
\alias{.glmm_G_FF}
\title{Diagonal of the fiber metric block G_FF at one parameter point}
\usage{
.glmm_G_FF(sigma, alpha, X, group, beta)
}
\arguments{
\item{sigma}{Scalar; group-level SD.}

\item{alpha}{J-vector; group intercepts.}

\item{X}{N x 2 matrix; fixed-effect design matrix.}

\item{group}{N-vector (integer 1..J); group membership.}

\item{beta}{2-vector; fixed-effect coefficients.}
}
\value{
J-vector (diagonal of G_FF).
}
\description{
G_FF\link{j,j} = 1/sigma^2  +  sum_{i in group_j} p_ij * (1 - p_ij)
}
\keyword{internal}
