% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_metric_glmm.R
\name{.glmm_grad_vec}
\alias{.glmm_grad_vec}
\title{Gradient of the log-posterior as a flat P-vector}
\usage{
.glmm_grad_vec(mu, log_sigma, alpha, beta, stan_data)
}
\value{
Numeric vector of length P, in order
(grad_mu, grad_log_sigma, grad_alpha[1:J], grad_beta[1:2]).
}
\description{
Gradient of the log-posterior as a flat P-vector
}
\keyword{internal}
