% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_posterior_glmm.R
\name{.state_to_list}
\alias{.state_to_list}
\title{Wrap a named parameter list into a state vector and back}
\usage{
.state_to_list(state, J)
}
\description{
Wrap a named parameter list into a state vector and back
}
\keyword{internal}
