% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_metric_glmm.R
\name{.unpack}
\alias{.unpack}
\title{Extract parameter components from a flat state vector}
\usage{
.unpack(q, J)
}
\description{
Extract parameter components from a flat state vector
}
\keyword{internal}
