% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{print.fibr_connection}
\alias{print.fibr_connection}
\title{Print a fibr_connection object}
\usage{
\method{print}{fibr_connection}(x, ...)
}
\arguments{
\item{x}{A \code{fibr_connection} object.}

\item{...}{Ignored.}
}
\description{
Print a fibr_connection object
}
