% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_holonomy.R
\name{print.fibr_holonomy}
\alias{print.fibr_holonomy}
\title{Print a fibr_holonomy object}
\usage{
\method{print}{fibr_holonomy}(x, ...)
}
\arguments{
\item{x}{A \code{fibr_holonomy} object.}

\item{...}{Ignored.}
}
\description{
Print a fibr_holonomy object
}
