% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvature.R
\name{print.fibr_transport}
\alias{print.fibr_transport}
\title{Print a fibr_transport object}
\usage{
\method{print}{fibr_transport}(x, ...)
}
\arguments{
\item{x}{A \code{fibr_transport} object.}

\item{...}{Ignored.}
}
\description{
Print a fibr_transport object
}
